requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
pytest>=8.2,<9.0
httpx>=0.27,<1.0
Pillow>=10.4,<12.0
